Split factory into several files
