+++ ttnn/cpp/ttnn/operations/reduction/generic/device/reduce_op_multi_core_w_program_factory.cpp
@@ -1,0 +1,214 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include "reduce_op_multi_core_w_program_factory.hpp"
+#include "ttnn/operations/reduction/generic/device/reduce_op.hpp"
+#include <tt-metalium/work_split.hpp>
+#include <tt-metalium/constants.hpp>
+#include <tt-metalium/host_api.hpp>
+#include <tt-metalium/tensor_accessor_args.hpp>
+#include <cmath>
+
+using namespace tt::constants;
+
+namespace ttnn::operations::reduction::generic::program {
+
+ReduceMultiCoreWProgramFactory::cached_program_t ReduceMultiCoreWProgramFactory::create(
+    const operation_attributes_t& operation_attributes,
+    const tensor_args_t& tensor_args,
+    ReduceTensorReturnValue& tensor_return_value) {
+    using namespace tt;
+    using namespace tt::tt_metal;
+    const auto& a = tensor_args.input_tensor;
+    auto& output = tensor_return_value;
+    const auto& shape = a.padded_shape();
+    uint32_t W = shape[3], H = shape[2], NC = shape[1] * shape[0];
+
+    uint32_t Wt = W / TILE_WIDTH;
+    uint32_t Ht = H / TILE_HEIGHT;
+
+    auto [math_fidelity, math_approx_mode, fp32_dest_acc_en, packer_l1_acc, dst_full_sync_en] =
+        get_compute_kernel_config_args(a.device()->arch(), operation_attributes.compute_kernel_config);
+
+    tt_metal::Program program = tt_metal::CreateProgram();
+
+    tt::DataFormat src0_cb_data_format = tt_metal::datatype_to_dataformat_converter(a.dtype());
+    uint32_t src0_single_tile_size = tt::tile_size(src0_cb_data_format);
+    tt::DataFormat scaler_cb_data_format = tt::DataFormat::Float16_b;
+    uint32_t scaler_single_tile_size = tt::tile_size(scaler_cb_data_format);
+    tt::DataFormat dst_cb_data_format = tt_metal::datatype_to_dataformat_converter(output.dtype());
+    uint32_t dst_single_tile_size = tt::tile_size(dst_cb_data_format);
+
+    tt_metal::IDevice* device = a.device();
+
+    auto compute_with_storage_grid_size = device->compute_with_storage_grid_size();
+    auto num_rows = NC * Ht;
+    uint32_t num_cores;
+    CoreRangeSet all_cores, core_group_1, core_group_2;
+    uint32_t num_rows_per_core_group_1, num_rows_per_core_group_2;
+    if (operation_attributes.sub_core_grids.has_value()) {
+        std::tie(
+            num_cores, all_cores, core_group_1, core_group_2, num_rows_per_core_group_1, num_rows_per_core_group_2) =
+            tt::tt_metal::split_work_to_cores(*operation_attributes.sub_core_grids, num_rows);
+    } else {
+        std::tie(
+            num_cores, all_cores, core_group_1, core_group_2, num_rows_per_core_group_1, num_rows_per_core_group_2) =
+            tt::tt_metal::split_work_to_cores(compute_with_storage_grid_size, num_rows);
+    }
+
+    uint32_t src0_cb_index = 0;
+    uint32_t num_input_tiles = 2;
+    tt_metal::CircularBufferConfig cb_src0_config =
+        tt_metal::CircularBufferConfig(num_input_tiles * src0_single_tile_size, {{src0_cb_index, src0_cb_data_format}})
+            .set_page_size(src0_cb_index, src0_single_tile_size);
+    tt_metal::CreateCircularBuffer(program, all_cores, cb_src0_config);
+
+    tt_metal::CircularBufferConfig cb_scaler_config =
+        tt_metal::CircularBufferConfig(
+            num_input_tiles * scaler_single_tile_size, {{CBIndex::c_2, scaler_cb_data_format}})
+            .set_page_size(CBIndex::c_2, scaler_single_tile_size);
+    tt_metal::CreateCircularBuffer(program, all_cores, cb_scaler_config);
+
+    uint32_t output_cb_index = tt::CBIndex::c_3;
+    uint32_t num_output_tiles = 2;
+    tt_metal::CircularBufferConfig cb_output_config =
+        tt_metal::CircularBufferConfig(num_output_tiles * dst_single_tile_size, {{output_cb_index, dst_cb_data_format}})
+            .set_page_size(output_cb_index, dst_single_tile_size);
+    tt_metal::CreateCircularBuffer(program, all_cores, cb_output_config);
+
+    bfloat16 bfloat_scaler_value = bfloat16::truncate(operation_attributes.scaler);
+    uint32_t packed_scaler_value = pack_two_bfloat16_into_uint32({bfloat_scaler_value, bfloat_scaler_value});
+    tt_metal::Buffer* src_buffer = a.buffer();
+    std::vector<uint32_t> reader_compile_time_args = {packed_scaler_value};
+    TensorAccessorArgs(*src_buffer).append_to(reader_compile_time_args);
+    tt_metal::Buffer* dst_buffer = output.buffer();
+    std::vector<uint32_t> writer_compile_time_args = {(std::uint32_t)output_cb_index};
+    TensorAccessorArgs(*dst_buffer).append_to(writer_compile_time_args);
+
+    std::map<std::string, std::string> reduce_defines =
+        reduce_op_utils::get_defines(operation_attributes.math_op, ReduceOpDim::W);
+    tt_metal::KernelHandle reader_kernel_id = tt_metal::CreateKernel(
+        program,
+        "ttnn/cpp/ttnn/operations/reduction/generic/device/kernels/dataflow/"
+        "reader_unary_reduce_universal_start_id.cpp",
+        all_cores,
+        tt_metal::ReaderDataMovementConfig(reader_compile_time_args, reduce_defines));
+
+    tt_metal::KernelHandle writer_kernel_id = tt_metal::CreateKernel(
+        program,
+        "ttnn/cpp/ttnn/operations/eltwise/unary/device/kernels/dataflow/writer_unary_universal_start_id.cpp",
+        all_cores,
+        tt_metal::WriterDataMovementConfig(writer_compile_time_args, reduce_defines));
+
+    std::vector<uint32_t> compute_kernel_args_group_1 = {
+        num_rows_per_core_group_1,  // Ht
+        Wt,                         // Wt
+        1,                          // NC
+    };
+
+    tt_metal::CreateKernel(
+        program,
+        "ttnn/cpp/ttnn/operations/reduction/generic/device/kernels/compute/reduce_w.cpp",
+        core_group_1,
+        tt_metal::ComputeConfig{
+            .math_fidelity = math_fidelity,
+            .fp32_dest_acc_en = fp32_dest_acc_en,
+            .compile_args = compute_kernel_args_group_1,
+            .defines = reduce_defines});
+
+    if (!core_group_2.ranges().empty()) {
+        std::vector<uint32_t> compute_kernel_args_group_2 = {
+            num_rows_per_core_group_2,  // Ht
+            Wt,                         // Wt
+            1,                          // NC
+        };
+
+        tt_metal::CreateKernel(
+            program,
+            "ttnn/cpp/ttnn/operations/reduction/generic/device/kernels/compute/reduce_w.cpp",
+            core_group_2,
+            tt_metal::ComputeConfig{
+                .math_fidelity = math_fidelity,
+                .fp32_dest_acc_en = fp32_dest_acc_en,
+                .compile_args = compute_kernel_args_group_2,
+                .defines = reduce_defines});
+    }
+
+    uint32_t out_dim_divider = Wt;
+    std::vector<CoreCoord> cores;
+    if (operation_attributes.sub_core_grids.has_value()) {
+        for (const auto& range : all_cores.ranges()) {
+            for (int y = range.start_coord.y; y <= range.end_coord.y; ++y) {
+                for (int x = range.start_coord.x; x <= range.end_coord.x; ++x) {
+                    cores.emplace_back(x, y);
+                }
+            }
+        }
+    } else {
+        cores = grid_to_cores(num_cores, compute_with_storage_grid_size.x, compute_with_storage_grid_size.y, false);
+    }
+    for (uint32_t i = 0, num_tiles_read = 0; i < num_cores; i++) {
+        const CoreCoord& core = cores[i];
+        uint32_t num_rows_per_core = 0;
+        if (core_group_1.contains(core)) {
+            num_rows_per_core = num_rows_per_core_group_1;
+        } else if (core_group_2.contains(core)) {
+            num_rows_per_core = num_rows_per_core_group_2;
+        } else {
+            TT_THROW("Core not in specified core ranges");
+        }
+        uint32_t num_tensor_tiles_per_core = num_rows_per_core * Wt;
+        tt_metal::SetRuntimeArgs(
+            program,
+            reader_kernel_id,
+            core,
+            {
+                a.buffer()->address(),
+                num_tensor_tiles_per_core,
+                num_tiles_read  // tile index of row to start reading from
+            });
+
+        tt_metal::SetRuntimeArgs(
+            program,
+            writer_kernel_id,
+            core,
+            {
+                output.buffer()->address(),
+                num_tensor_tiles_per_core / out_dim_divider,  // number of tiles to write
+                num_tiles_read / out_dim_divider              // output tile start index
+            });
+        num_tiles_read += num_tensor_tiles_per_core;
+    }
+
+    return {std::move(program), {reader_kernel_id, writer_kernel_id, cores}};
+}
+
+void ReduceMultiCoreWProgramFactory::override_runtime_arguments(
+    cached_program_t& cached_program,
+    const operation_attributes_t& operation_attributes,
+    const tensor_args_t& tensor_args,
+    ReduceTensorReturnValue& tensor_return_value) {
+    using namespace tt;
+    using namespace tt::tt_metal;
+    auto* src_dram_buffer = tensor_args.input_tensor.buffer();
+    auto* dst_dram_buffer = tensor_return_value.buffer();
+
+    auto& reader_runtime_args_by_core =
+        GetRuntimeArgs(cached_program.program, cached_program.shared_variables.reader_kernel_id);
+    auto& writer_runtime_args_by_core =
+        GetRuntimeArgs(cached_program.program, cached_program.shared_variables.writer_kernel_id);
+    for (const auto& core : cached_program.shared_variables.cores) {
+        {
+            auto& runtime_args = reader_runtime_args_by_core[core.x][core.y];
+            runtime_args[0] = src_dram_buffer->address();
+        }
+
+        {
+            auto& runtime_args = writer_runtime_args_by_core[core.x][core.y];
+            runtime_args[0] = dst_dram_buffer->address();
+        }
+    }
+}
+
+}  // namespace ttnn::operations::reduction::generic::program
